requirements.txt
openai
python-dotenv
pydantic
requests
click
kiutils
jinja2
sqlite3-api
